+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes in keycloak-group-management will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+### Added 
+- Admins can directly add new members to groups
+
 ## [1.4.1]- 2025-02-03
 
 ### Fixed